feat: initial curric login interim solution for docx downloads
